moved run test to test_main.py
